requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
facenet-pytorch==2.6.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
grpcio==1.66.1
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxconverter-common==1.14.0
onnxruntime==1.19.2
onnxscript==0.1.0.dev20240909
packaging==24.1
pillow==10.2.0
protobuf==3.20.2
requests==2.32.3
six==1.16.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
